lab5/lab4/requirements.txt
Django==3.1.7
psycopg2-binary>=2.8
redis>=3.5.3
django-rest-framework==0.1.0
djangorestframework==3.12.4

